Add files via upload
allEncompassingPlot will plot a single very data filled, comprehensive color coated bar graph with additional boxes of top user stats and 'GOAT' conversation for the 3 users with the best overall stats. 
specificStats will plot 5 separate graphs. # of posts, # of likes given, # of likes received, avg likes/post for infrequent posters (<10 posts), and avg likes/post for frequent posters (>10 posts).
The other files are unimportant. avgLikesPerPost plots what it says in the name. Names 2IDs prints names and their IDs. sandBox was just for me to play around with things. usingConv&Every was my first attempt at cleaning the data with the conversation.json and everybody.json files that the groupme export provides. The data was easier to clean in those jsons, but it was insufficient. allEncompassingPlot uses message.json which proved to be way better despite being harder to clean.